DiscoveryWeekTeam4/ifel tower vs pissa tower/requirement.txt
streamlit
numpy
opencv-python
Pillow
keras
tensorflow
streamlit-webrtc
av
